Copyable Spawnable Tickets.
Spawnable Tickets can now be copied. Behind the scenes a ref counted object is used to track the status of the tickets so they get cleaned when there are no more copies of a ticket in use. To support this the call to retrieve a ticket instance from a ticket id had to be updated to return a copy, which had the additional side-effect that the mutex around the ticket id to ticket map is no longer needed. This change also allows the spawnable asset associated with the ticket to be retrieved.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>