Unifying drag and drop behavior. Dragging from asset browser to the v… (#7858)
* Unifying drag and drop behavior. Dragging from asset browser to the viewport and outliner should behave the same now when you have multiple products selected.
The primary place this was an issue was with FBX files that had multiple products.

Before this change, you selected the source FBX and dragged it to the outliner you would get one entity with a component per product. If you dragged it to the viewport, you got an entity for each product. Now both match the behavior of the outliner. You get one entity with a component per product.

Note that this does not fix bugs or unexpected behavior with this. If the dragged FBX has multiple products, it will still create conflicting components and only populate the first one. There's a separate initiative to define best practice in this situation. The goal here was to make the behavior the same when dragging from the asset browser to the outliner or viewport.

Signed-off-by: AMZN-stankowi <4838196+AMZN-stankowi@users.noreply.github.com>

* Removed stale comment

Signed-off-by: AMZN-stankowi <4838196+AMZN-stankowi@users.noreply.github.com>