[Linux] Respond to the `_NET_WM_PING` client message event
Some window managers (like mutter) will query an application at some
interval to see if it is processing its event loop. They do this by sending
the window a `_NET_WM_PING` message. The application is expected to process
that message and send it back to the server.

During normal execution, responding to the `_NET_WM_PING` message is
handled by Qt. However, when the user enters game mode in the editor,
processing of Qt's event loop stops, and everything is processed by the
XcbNativeWindow instead. This adds the proper response to that event, so
that mutter does not consider the Editor process as hung while the user is
in game mode.

Signed-off-by: Chris Burel <burelc@amazon.com>