[Linux] Do not filter XCB_CLIENT_MESSAGE events when in game mode in the Editor
XCB_CLIENT_MESSAGE events include _NET_WM_PING events, which window managers
use to detect if an application is still responding. When Qt creates the
window, it sets the _NET_WM_PING atom of the WM_PROTOCOLS property, so window
managers will expect the application to support this protocol. By skipping the
filtering of this event, Qt processes the ping event normally, so that window
managers do not think that the Editor has stopped responding.

Signed-off-by: Chris Burel <burelc@amazon.com>