Adding support for Gem Filesystem aliases.
This allows gems to portability reference its source directory in C++.

For the FileIO system using a key with the pattern of`@gemroot:<gemname>@` attempts to resolve to the gem root folder with that gemname
For the AssetProcessor settings the `@GEMROOT:<gemname>@` alias can be used to reference the root of the gem.

Updated the cmake_dependencies.*.setreg file to contain objects of the actual Gem names as stored in the gem.json "gem_name".

Previously the cmake_dependencies.*.setreg only contained the name of Gem module targets from cmake
The new format of the cmake_dependecies.*.setreg is as follows
```
{
    "O3DE": {
        "Gems": {
            "<actual gem_name>": {
                "Targets": {
                    "<gem module cmake target 1>" : {
                        ...
                    },
                    "<gem module cmake target N>" : {
                        ...
                    }
                },
            }
        }
    }
}
```

The Root Gem paths are now merged to the Settings Registry via visiting
each of the registered o3de manifest files.
- o3de_manifest.json
- project.json
- engine.json

These manifest files are recursed through for each
"external_subdirectories" entry found within them.

Any gem.json files found within an external subdirectory entry also has
its "external_subdirectories" entries walked.

Each external_subdirectory path which exists on the filesystem and
contains a gem.json file is added to a section within the Settings
Registry for mapping the gem name to gem path on disk.
```
{
    "O3DE": {
        "Manifest": {
            "Gems": {
                "Atom": {
                    "Path": "D:/o3de/Gems/Atom"
                },
                "Atom": {
                    "Path": "D:/o3de/Gems/EMotionFX"
                },
                "TestGem": {
                    "Path": "D:/TestGem"
                },
                "TestProject": {
                    "Path": "D:/TestProject/Gem/"
                }
            }
        }
    }
}
```

Active Gems are located under the key of "/O3DE/Gems"
Registry Manifest Gems are located under the key  of
"/O3DE/Manifest/Gems"

The path of Active Gems are determined by querying the Gem Name for each
gem by visiting each field name of "/O3DE/Gems/<GemName>" in the Active
Gem section.
Afterwards, that Gem Name is used as a key into the
"/O3DE/Manifest/Gems/<GemName>" section of the Settings Registry to find
the settings object which contains the "Path" key.

The value of the Path Key is the absolute path to the Gem root folder.

Additional Changes:
Added support for an o3de-manifest-path option to specify the directory
where the o3de_manifest.json file is located at.
The only use case for this is UnitTest.
Non-test code shouldn't use this option, as it can affect locating the
engine root and gems root directories, since the project has to cross check
the o3de_manifest.json to locate the paths.

fixes #8104

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>