Add fragmentation and summarized RHI pool stats
This commit modifies the GPU memory statistics reporting mechanism to
include fragmentation information, added as a new field to the
HeapMemoryUsage. Fragmentation is reported as a normalized value from 0
to 1, with 0 representing no fragmentation. Buffer memory statistics
also have the ability to report fragmentation in cases where buffers
themselves are fragmented, although this functionality is not yet used.

Fragmentation itself is implemented only for the FreeListAllocator which
measures fragmentation according to the following formula:

    1 - [largest free block size] / [unused bytes]

There are many measures of fragmentation available (see:
https://www.cs.tufts.edu/~nr/cs257/archive/paul-wilson/fragmentation.pdf),
but this method was chosen for efficiency and simplicity. A test was
added to verify this measure operates as expected.

In addition, this change modifies the GPU memory profiler view to
include a summary table of each resource pool, displaying budget (if
specified), reserved memory, resident memory, and pool fragmentation.
This table is independently displayed and sorted relative to the
existing table.

Finally, this change modifies the Atom_Utils library slightly. Before,
the ImGui GPU memory profiler was contained entirely within a header,
meaning that users needed to ensure they linked a specific set of
dependencies. This change moves that code to a separate translation unit
within the utils library and fixes broken dependencies as well.

Signed-off-by: Jeremy Ong <jcong@amazon.com>