Fix duplicate wildcard dependencies being saved to database
When trying to resolve unresolved dependencies, existing dependencies weren't taken into account and the query to save them was just inserting without doing any checks.
Added a new query to get all the dependencies for a given product.
PathDependencyManager will now grab the list of existing dependencies and remove existing ones from the list of to-be-saved dependencies

Signed-off-by: amzn-mike <80125227+amzn-mike@users.noreply.github.com>