Updated references within Gems being within the Engine root folder.
The references were updated to query the Settings Registry via the "/O3DE/Manifest/Gems/<GemName>/path" key to locate the Gem root path.
FileIO references to paths such as "@engroot@/Gem/<GemName>", were updated to be "@gemroot:<GemName>@"

Added a Test Utility function AZ::Test::AddActiveGem to a Gem name
object to the "/O3DE/Gems/<GemName>" section(which represents the Active
Gem list) and to query the GemPath from the manifest and add a
Filesystem @gemroot@ alias for the Gem.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>