Update Freetype to 2.11.1, fix DLL_IMPORT usage (#8069)
* Update Freetype to 2.11.1, fix DLL_IMPORT usage

This change updates freetype2 to use v2.11.1, the latest stable
release of freetype2 at this time.  As part of this, it was noticed
that there were still legacy (CryEngine) uses of a macro called
DLL_IMPORT and DLL_EXPORT.  These macros are deprecated and were
simply forwardng to AZ_DLL_IMPORT and AZ_DLL_EXPORT.

Roughtly 8 instances where code was using the old DLL_IMPORT instead of
AZ_DLL_IMPORT has been switched to use AZ_DLL_IMPORT instead, likewise
with export macros, and DLL_IMPORT is no longer defined.

This is a necessary change because some 3rd party libraries (freetype2
is not the only one) coincidently use DLL_IMPORT or DLL_EXPORT as a
define to control whether they are compiled in DLL mode or static
library mode - those symbols are thus dangerous to define at global
scope in any project that uses these libraries as it will cause
them to be confused between when they are compiled (as a static lib)
but then included into an engine which defines these macros, and will
do things like expect a dll version of their library when it was
compiled statically.