Fixed ly_create_alias to propagate link dependencies when multiple dependent targets are specified.
Moved the ly_create_alias calls for the "Atom" gem  and "AtomLyIntegration" gem to their top level
CMakeLists.txt in order to properly determine the Gem root for the
"Atom" gem.

Moved the populate of the GemName to GemPath map in the
PlatformConfiguration::ReadRecognizersFromSettingsRegistry outside of
the ScanFolder loop so that it occurs once.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>