[development] prevent the move operation in AZStd::vector::erase when the start/end iterators are the same (#8316)
The introduction of #5735 exposed a case where undefined behaviour was being allowed to occur in the range based version of AZStd::vector::erase. This was discovered in the Profiler gem ImGui visualizer through the use of AZStd::upper_bound not finding the desired value and returning begin() thus producing vector::erase(begin(), begin()). Rather than patching that instance (#8306), this change is to fix the root issue.

Signed-off-by: AMZN-ScottR 24445312+AMZN-ScottR@users.noreply.github.com