Fix runtime shape drawing in Editor. (#8325)
If runtime shape components with debug drawing are spawned in the Editor (for example thru the dynamic vegetation system), they weren't previously being drawn. This was due to the fact that debug drawing in the Editor outside of game mode in controlled through Editor entity visibility, which queries the EditorEntityInfo to see if the entity is visible. Runtime entities don't have EditorEntityInfo, so they were invisible by default. This change makes them visible but not selectable.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>