[Linux] Improve thread name support (#6704)
pthread only supports a thread name up to 15 characters + null terminator.
Consequently, if the user code was providing a name that was too long, it
would get the default "AZStd::thread" name. This change uses an elided
name, removing characters from the middle of the name, so that it fits
within the 16 byte limit. The decision to elide the middle was made so that
threads can be named with trailing numbers, and the number will be retained
in the thread's name.

Signed-off-by: Chris Burel <burelc@amazon.com>