Updates OpenEXR to rev4 and fixes dependency bugs
* Updates OpenEXR to Revision 4
* Fixes a bug where missing quotes in our install scripts would
  not correctly interpret a list of target names
* Fixes a bug in the dependency walker function:

The dependency walker function inside RuntimeDependencies_common
assumed that mapped configurations (MAP_IMPORTED_CONFIG_xxxxx) on a
target was a single value like "RELEASE" but it is often a list.

This simplifies and fixes that code so that if a list is supplied
it still functions.  In addition, by default, if no mapping is
specified, we will first try to use IMPORTED_LOCATION_xxxxx where xxxxx
is the one that matches the same config name, and fall back to the
blank default one if not found.

So for example, if no mapping is specified for a given target and
we are in the PROFILE configuration, we'll prefer to use
IMPORTED_LOCATION_PROFILE if present, and if not, fallback to just
IMPORTED_LOCATION.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>