Fixed a bug in MaterialComponent that prevented UV overrides from being applied correctly in the viewport.
Repro steps:
Create a new level.
Select the shaderball entity in the default scene.
In the material inspector on the Mesh Component, click "Add Material Component".
On the "shaderball_01" slot, click "Edit Material Instance UV Map..."
Change the "Tiled" to "UV_packed". Notice the texture will change to be mapped differently in the viewport.
Change the it to "UV_world_1m". Notice that nothing changes in the viewport.
Change it back to "[Same as in the material]" which was the original value. Notice that still nothing changes. It should have changed back to the way it looked before.

It only worked the first time because of the LoadMaterials() call. Any subsequent change needs QueueMaterialUpdateNotification() to be called.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>