Updated the Linux Clang compiler detection to first search for "clang-<version>" followed "clang"
Removed from the project templates the compiler detection cmake files.
When using a Project-centric workflow now, the CompilerSettings.cmake script will be included before the cmake `project()` command through the Findo3de.cmake script that is part of the Engine.
This removes the need for having the project templates CompilerSettings.cmake and CompilerSettings_linux.cmake in the first place.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>