Add EBus support for shared_mutex (#8758)
* Add EBus support for shared_mutex.
EBus currently supports LocklessDispatch, which enables the dispatch to run without *any* locks, meaning that it doesn't provide any safety if BusConnect / BusDisconnect can occur in parallel with the dispatches. However, it doesn't currently provide a mechanism to allow for parallel dispatches that still block vs BusConnect / BusDisconnect. The theoretical solution is to use an EBus MutexType of shared_mutex, but this doesn't quite work in practice since there are some nuances required to handle the different types of locking and nested locking, and there's no one single solution that necessarily meets all needs.

The solution in this PR is to introduce EBusSharedDispatchMutex as a new optional custom MutexType that implements a type of shared_mutex support, where multiple (and optionally nested) dispatches can occur in parallel, but they will exclusively block with BusConnect / BusDisconnect. The EBus code needed to be modified minimally to expose more custom LockGuard types that can be overridden, so that it's possible to get different locking semantics based on the type of locking that the EBus is trying to do.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Addressed PR feedback
* Removed % from EBusSharedDispatchMutex in comments
* Added EBusSharedDispatchTraits<Bus> for easier usage
* Renamed header file to EBusSharedDispatchTraits.h to reflect the new intended usage pattern
* Changed unit test to use the new the EBusSharedDispatchTraits
* Removed AZ_FORCE_INLINE
* Added more comments describing how the DispatchLockGuard works
* Removed unused m_ownSharedLockOnThread variable
* Removed "using namespace AZ"
* Removed unused Connect policy
* Fixed comment above the Disconnect policy

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Addressed PR feedback
* Added acquire() call so that there is a balance of acquire() and release() in every unit test.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>