Resolve AZ::IO::Streamer's RequestPath upon construction.
Resolving relative paths and aliases to absolute paths is now done when a RequestPath is constructed instead of using lazy initialization. Practically this means that AZ::IO::Streamer rarely has to resolve paths and the code calling into Streamer will instead resolve the path. Overall this shaved off about 5 to 10 microseconds and reduced the gap between the drive read speed and the overall read speed with Streamer by 1-2Mib/s (on a SATA SSD).

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>