Adding CMakePresets support for O3DE.
The host platforms of Windows, Linux and MacOS has added support for
configure, build and test presets.
The non-host platforms of iOS and Android has supported for configure
and build presets.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

Filled out the CMakePresets configuration, build and test presets for all platforms.

A monolithic and non-monolithic preset has been added to each platform along with a unity/no-unity build preset and presets for the de-facto build generators for each platform.

The host platforms in addition have build presets for the Editor, AssetProcessor and the test suites of smoke and main.
Furthermore a test preset has been addded to run those test.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>