fix sc display on window open & interpreter crashes
SC Bug fixes, in response to PR feedback and QA testing

Fix failure to display bugs when a new window or tab is opened by
clicking the open graph shortcut button in the editor. This was done by
enforcing that the graph, whether or not previously opened, was
selected as the active asset.

Added parser errors for direct entity references . Added proper EntityId
value translation for invalid Ids.

Signed-off-by: carlitosan <82187351+carlitosan@users.noreply.github.com>