Graph Model: Added more numeric and math types to json serializer
Adding basic math and numeric types, needed by material canvas, to graph model slot serializer.
This does not include advanced types like assets and other things that will be needed in the near future.
Eventually this should be pluggable so that new types can be added or natively supported by the serializer.

Signed-off-by: Guthrie Adams <guthadam@amazon.com>