Fixed physx material ids being set to null when used in a prefab (#8924)
The `OnReadEnd` function of the serialization of `MaterialSelection` class tries to validate the material ids after serialization by looking into the material library. The problem was when executed in the context of Asset Builder (due to a prefab), the material library is never loaded by the physics system because it's expecting the event "CriticalAssetsCompiled", which is not fired at asset processor time.

Fixed by adding a protection in `OnReadEnd` to only validate the material ids if it's not an asset builder. That way when prefabs are processing the data it will only write the ids out, and during editor/launcher it will validate it against the library.

Fixes #7550 

Signed-off-by: moraaar <moraaar@amazon.com>