Added thread safety to shape components (#8948)
* Added thread safety to shape components.
* Created unit tests that demonstrated threading issues with current components
* Added mutexes to allow for concurrent queries but blocking data changes
* Modified shape request bus to allow for concurrent queries

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Addressed PR feedback

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Removed constexpr from lambda capture.
This appears to cause warnings/errors on some platforms.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Apparently constexpr float needs to be lambda-captured, and constexpr int needs to be NOT captured.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Fixed bug when using Shape Reference with Vegetation Spawner.
I previously removed the OnShapeChanged notification on Activate(), thinking that there wasn't a situation in which it could be received. However, there is a specific case with the Vegetation Spawner and editing the Shape Reference in the Editor that requires this event to be fired from activation. I've re-added the notification along with a comment to describe the reason it's needed.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Switched from constexpr to const.
Simplifies compliance with lambda capture warnings across compilers.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Safer cross-compiler answer to lambda captures.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* PR feedback - prefer "using"

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Fix test name.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Fixed bug exposed by rearranging BusConnect order.
The PolygonPrismShape was keeping stale callback pointers around, and these stale callbacks were called as a part of the Activate() call. Due to the rearranged bus connection order, this caused some physics tests to fail because the callbacks were executing logic that called back to the PolygonPrismShape even though it hadn't fully activated yet. This fixes the bug by clearing the pointer, and adds a unit test that failed prior to any code changes and succeeds with the bugfix.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Addressed PR feedback.
Reduced the number of iterations on the thread safety tests to 30,000 which is 2X the minimum number needed to reliably cause errors locally.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>