Make Uuid / AssetId ToFixedString calls more consistent and less error prone.  (#9020)
* add ToFixedString() methods to Uuid and AssetId

Using the same ToString<>() methods with the Uuid MaxStringBuffer leads
to an easy to fallint trap and error, that frequently only reveals
itself when some error condition with var args is hit. This probably
won't happen during initial testing. This should give an easy to find
safer option.

Signed-off-by: carlitosan <82187351+carlitosan@users.noreply.github.com>

* consolidate ToFixedString functionality for Uuid and AssetId

Both AssetId and Uuid have template functionality to allow AZStd::fixed_string as
the template instatation of their ToString calls. However, this leaves an
easy trap to fall into, especially in string format calls. It is easy to
use the Uuid::MaxStringBuffer when calling the AssetId to fixed string calls,
which leads to errors since the AssetId requires extra characters for the subId.

This change adds ToFixedString member functions to both classes, and updates
all direct references in the codebase to Uuid::MaxStringBuffer to use AssetId.

Signed-off-by: carlitosan <82187351+carlitosan@users.noreply.github.com>