[Terrain] Fix shape area falloff to calculate either 2D or 3D falloff. (#9057)
* Fix shape area falloff to only calculate 2D falloff, not 3D.
The previous approach was using the height of the query point as a part of the distance calculating, which gave us full 3D falloff. This really wasn't desirable behavior for either vegetation or terrain, so the gradient has been switched to use 2D falloff instead. We might want to consider eventually adding support for the Gradient Transform and making 2D / 3D an optional choice.
Added unit tests to validate and enforce the new behavior.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Added 3D Falloff toggle.
There are use cases where either 2D falloff or 3D falloff are useful, so changed this to be an optional toggle.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>