Bug fix for asset appearing multiple times in material editor asset selectors
This addresses an issue where the asset grid or combo box could potentially have the same asset added multiple times.
This will not address the case if an asset in a different folder has the same file name.
That will result in the same name appearing in the list or dropdown multiple times but they will point to the correct asset.

Signed-off-by: Guthrie Adams <guthadam@amazon.com>