Fix memory leak when storing the generic class info for Assets
Asset types are serialized with a custom serializer that uses the generic
serialization API. This allows for the serializer to be templated on the
asset type. Normally, the serialize context tries to create those generic
serializers only once, by keeping a map of AZ::TypeId->Serializer*. It
would look in the map to see if a serializer was already present, and only
if it wasn't found, it would create the serializer for that type.

However, the GenericClassInfo is allowed to override the typeid that
should be used to identify the types it handles. The map lookup was done
with the "canonical" typeid (ie, the result of AzTypeInfo<T>::Uuid()),
but then stored based on the generic's specialized typeid, (ie, the
result of SerializeGenericTypeInfo<T>::ClassInfoType::GetSpecializedTypeId()).
Consequently the lookup for a serializer of any asset type would never find
a previously made serializer. To make matters worse, the code that inserts
into the map does not check for pre-existing values, so it would overwrite
the existing Serializer*, resulting in a memory leak.

This fixes that issue by having the GenericTypeInfo for Assets return
the real specialized typeid for the asset type in the
`GetSpecializedTypeId()` method. Then The use the actual derived
Data::Asset<Data::AssetType>> typeid to query the Generic Asset ClassId.
That value will then be compared against GetAssetClassId() in the
serialization code.

Signed-off-by: Chris Burel <burelc@amazon.com>