Added new PhysX material asset where 1 asset is 1 material (#330)
The properties of the physx materials are the same, the difference is that the asset is now the material, not a material library, which simplifies the code and makes it work the same as render materials. 

- **MaterialConfiguration**: structure with physx properties.
- **MaterialAsset**: asset data that has a MaterialConfiguraiton. This asset will enable Asset Editor to create this assets.
- **Material**: runtime material with the physx internal classes, created from a MaterialConfiguration.

Asset Editor will create physx materials files (`.physxmaterial`) in the project's asset folder, then an Asset Processor rule copies the file to the cache folder, which makes it ready for O3DE to use. The old asset library extension is `.physmaterial`.

Signed-off-by: moraaar <moraaar@amazon.com>