Updated the EnvironmentVariable class boolean checks.
The `operator bool` and `operator!` now check if the EnvironmentVariableHolderType type is non-nullptr and is in a constructed state.

This makes it consistent with the `operator*` and `operator->` methods and allows the construct as follows to always work
```
if (environmentVariable)
{
   auto value = *environmentVariable
}
```

Also removed the `Construct` environment variable class as it allows invoking the constructor over an already constructed value stored in the EnvironmentVariableHolder.

The `Construct` function itself required the EnvironmentVariable to be in a very specific state where the EnvironmentVariable has been constructed before and the destroyed in order to use safely.

The function that should be used to re-create an environment variable is the `AZ::Environment::CreateVariable` function.

fixes #9274

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>