Removed unnecessary PDO-specific depth bias. This bias was causing incorrect results in AtomSampleViewer's parallax screenshot tests. The bias was too large, and probably inconsisntent because it didn't take the slope of the geometry into account. Shadow acne can be addressed by adjusting the light's shadow bias and normal shadow bias. Maybe in the past one or both of these didn't work on PDO surfaces, but now they do.
Testing:
I set up a test scene with a cloth material using parallax PDO, and a spotlight, per the old bug report record associated with this issue. When I removed the PdoDepthBias, the acne appeared, and then when I adjuste the spotlight's bias and/or normal bias the acne was successfully resolved.
Ran the AtomSampleViewer parallax test script and visiuall confirmed that the shadows have returned, matching closer to the original expected screenshots. I will update the ASV screenshots with new versions using my latest results.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>