[EMotionFX] Fix use-after-free of Actor in PreviewMotionFixture.MotionCanBePlayed
This test uses a `unique_ptr<Actor>` to manage the lifetime of the Actor
instance, so the Actor is destroyed when the body of the test finishes.
But the ActorInstance maintains a pointer to the Actor, and it was not
being destroyed until the MainWindow was closed, which calls
`ClearScene()`. The RemoveActorInstance command that drives the scene
clearing calls `actorInstance->GetActor()->GetId()`, which was reading from
an already destroyed variable.

Signed-off-by: Chris Burel <burelc@amazon.com>