[EMotionFX] Fix tests that have use-after-free violations on Actor instances
Some EMotionFX tests use a `unique_ptr` to hold an Actor instance in the
body of the test, and then use that Actor* to create an ActorInstance
instance, but don't explicitly destroy the ActorInstance. Consequently,
when the test's teardown runs and clears the EMotionFX scene, the
ActorInstance has a pointer to an already deleted Actor. This prevents that
issue by adding the explicit commands to clean up the ActorInstances before
the body of the test exits.

Signed-off-by: Chris Burel <burelc@amazon.com>