[Android] Fix some serialization, networking, and shader compilation issues (#9311)
* [Android] Fix deserialization errors for assets stored as object stream.
On Android and Linux, int64_t and unit64_t map to long and unsigned
long. This is different from other platforms. The AZ::s64 and AZ::u64
types are mapped to the correct type for 64-bit integrals based on
platform. Using int64_t and uint64_t results in the serialization system
using different uuid on different platforms. Android assets are built on
Windows. So an asset using one of those types that was serialized on
Windows will not deserialize correctly on Android.

Signed-off-by: amzn-sj <srikkant@amazon.com>

* [Android] Call to epoll_wait causes the main thread to never wake up unless broken into on a debugger and resumed again. This ends up being required every frame.

Signed-off-by: amzn-sj <srikkant@amazon.com>

* The unique-idx parameter is needed on Android as well because we need to bind resources to unique register indices on Android. The namespace parameter is required for graphics library specific attributes.

Signed-off-by: amzn-sj <srikkant@amazon.com>

* Add comment explaing issue

Signed-off-by: amzn-sj <srikkant@amazon.com>

* Change DrawItemSortKey to AZ::s64 instead of just fixing the serialization issue.

Signed-off-by: amzn-sj <srikkant@amazon.com>

* Fix Linux build error

Signed-off-by: amzn-sj <srikkant@amazon.com>

* Fix more Linux build errors

Signed-off-by: amzn-sj <srikkant@amazon.com>

* Fix failing Linux tests

Signed-off-by: amzn-sj <srikkant@amazon.com>