[AzCore] Remove redundant tests that leak memory
These test cases are covered by the 3 tests immediately above this line.
Whether the function under test is called with a string literal, which
converts to a `const char*`, or the result of `strdup()` on that same
string literal, makes no difference to the function. Anstrdup()` then we
need an accompanying call to `free()`, so the simplest thing to do is
just remove these redundant tests.

Signed-off-by: Chris Burel <burelc@amazon.com>