Change the way terrain geometry is rendered to reduce draws (#9452)
Previously terrain was rendered using a single uniform grid of meshes laid out inside the world bounds. The meshes used standard LODs so further away meshes had fewer triangles, but were the same physical size. This worked for the short term but had some long term problems

Meshes that were far away had very few triangles per draw which is quite inefficient.
Number of meshes grew linearly with terrain area, which makes very large areas of terrain impossible.
Rendering large terrains over long distances required many thousands of draws.
The new mesh rendering system uses a fixed number of meshes (based on user settings) which are laid out around the camera using multiple grids of meshes of different resolutions. All the meshes contain the same number of vertices, but cover different amounts of space in the world depending on which LOD grid they're part of. The user specifies how many meters they want to see the terrain at highest resolution (how far the first LOD should cover) and also how far they want to see the terrain.

So for instance, if the first LOD needs to go out to 128 meters, and the total rendering distance is 4096 meters, then there are 6 grids created for various LODs - 128, 256, 512, 1024, 2048, and 4096. Each grid will have the same number of meshes and each mesh will have the same number of vertices. The meshes are stored internally with toroidal addressing, so as the camera moves around the world, typically only the edge meshes will need to be updated.

When rendering, meshes will be drawn if they overlap their LOD range at all, which means there will be some overlap where meshes of multiple LODs will be drawn. To avoid overdraw, the shader takes care of clipping triangles which aren't in the correct range. This leads to some occasional holes in the geometry along LOD boundaries but will be fixed in the future using CLOD (basically adjusting the higher resolution mesh to look like the lower resolution LOD as the LOD boundary approaches)

The primary effect of this PR is a significant reduction in the number of draws for terrain, and decoupling the terrain bounds from rendering. It is now possible to see terrain at much further distances, and the draw counts have been reduced by 80-90%.

Signed-off-by: Ken Pruiksma <pruiksma@amazon.com>