AZ::IO::Streamer now issues read requests through the task system.
Issuing read request with the OS was taking up the majority of the read time (when no files needed to be opened). By moving the actual call to the OS to the task system the scheduler can work on the next request in the meantime and allows issuing read requests to the OS to overlap. In the test scenario this marginally improved the read speed of the drive, but closed gap between the overall throughput of Streamer and the drive to less than 1%.
While the test scenario didn't provide a significant improvement, speeds running in developer mode (--cl_streamerDevMode=true) have more than quadrupled. It should be noted that in test case the same few files get repeatedly read from the disk, which means in dev mode the Windows file cache is mostly used, but this doesn't translate to similar performance in real world scenarios as files are not frequently loaded within a time frame that the file cache would be beneficial (especially with the Asset Manager's aggressive caching of assets). It also doesn't perfectly demonstrates the ability of NVMe drives as the read request calls to the OS are much longer than normal due to copying from the cache. It does however show a significant improvement in handling fast reads, demonstrating a better chance of Streamer keeping up with faster driver.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>