Fixed an edge case with shader hot reloading where a shader variant might not get loaded.
I am working on improvments to AtomSampleViewer's MaterialHotReloadTest sample so that it reports the actual shader variant being used rather than the variant being requested by the material. This more accurate reporting revealed some failures that were not detected before. When a ShaderVariantAsset was loaded by the ShaderVariantAsyncLoader, and then removed due to changes in the shader variant list, and then added back again, the re-added shader variant asset was never reported via AssetBus::OnAssetReady. This is because the asset was removed from AssetManager::m_assets so when AssetCatalog::AssetChanged called AssetManager::ReloadAsset, there was no asset present. The solution was to ensure that the contents of ShaderVariantAsyncLoader match the contents of the AssetManager; if it's not in the AssetManager then remove it from the ShaderVariantAsyncLoader. This triggers the necessary code paths to get the ShaderVariantAsset loaded again via ShaderVariantAsyncLoader::QueueLoadShaderVariantAsset.

I also renamed the ShaderAsset::GetVariant and GetRootVariant functions to GetVariantAsset and GetRootVariantAsset for clarity.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>