[EMotionFX] Fix crash from Inspector window when closing the EMotionFX editor
This fixes a crash introduced by
20873d8427f3f13ff53c6eab7140d4c1a7293726. The Inspector window reuses
the widgets it creates. It does this by calling `takeWidget()` paired
with `setWidget()` on its `QScrollArea`. Calling `takeWidget()` removes
the current widget's parent, and returns ownership back to the caller.
To ensure that both the content widget and the no selection widget are
both destroyed, the Inspector window now takes back ownership of
whatever widget is in the scroll area (which can only be one of the
content widget or the no selection widget), and then deletes both of
them.

Signed-off-by: Chris Burel <burelc@amazon.com>