Updates android to OpenSSL-1.1.1o and fixes a compile error
Android update to OpenSSL-1.1.1o - the previous update failed
becuase the build farm was on an older version of the NDK than
the package was built with.   Once the build farm is upgraded,
packages may use the new NDK, but until then, packages will use
the same NDK that the build farm does.

This also fixes a compile error (due to a warning-as-errors mode
triggering on a variable that was assigned but never used after
assignment).

See the openssl-1.1.1o-android-ndk-version-fix branch in
https://github.com/o3de/3p-package-source for the corresponding
changes to the package itself.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>