Update bootstrap.py
I've added double instead of single quotes for the docstrings in a few of these- this will eventually need to be augmented with proper formatting for Doxygen compatible docstrings, but that can come later. Also, at some point we will need to remove any code that is commented out. Also, many instances of "bootstrap" being misspelled.... throughout the DCCsi the first "t" seems to be missing. Wondering if you might have done a giant "find a replace" with the t missing?

Signed-off-by: benblack75 <benblac@amazon.com>
Signed-off-by: shawstar <shawstar@amazon.com>