Removed ConvertSystemSettings, ConvertSystemComponents and ConvertModuleComponents functions
As there are no existing Editor.xml/Game.xml files in O3DE, there is no reason to keep these conversion functions around.

Added a `dumptypes` command to the SerializeContextTools that can dump the reflected types(name and Uuid) to stdout or to a file using the --output-file.
The name and Uuid are separated by whitespace.

Added a `createtyoe` command to the SerializeContextTools that can create an instance of a reflected type and output it in json format. The instance can be written to stdout or to a file.

resolves #7567

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>