Fixed issue with AP waiting forever for builder registration
If the builder takes too long, the BuilderManager will timeout the process and terminate it.  AP will now react to a builder failure by printing an error message and shutting down.
Additionally, timeout was increased 3x from 5 minutes to 15 minutes as this situation was encountered when CPU was under heavy load.

Fixed issue with StartupErrorCollector crashing on shutdown when QuitRequested is called

Signed-off-by: amzn-mike <80125227+amzn-mike@users.noreply.github.com>