[Terrain] Fix and re-enable Terrain physics python test (#9858)
* Improve sketchy terrain automated test.
The test was creating a rigid body sphere that intersected the heightfield, dropped it, and tried to detect a collision. If the heightfield takes non-zero time to initialize (which it now does, because it's async), the ball would drop through the heightfield before the collision could be detected. Instead, it is now being created as a collider without a rigid body, so simulation timing is far less of a factor.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Removed for real this time.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Re-enabled and fixed the terrain physics test.
The test is changed to disable gravity, so that it simply detects a collision with a stationary sphere that's intersecting the heightfield.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Fixed race condition exposed by terrain test.
In rare cicrumstances, if the terrain system is shutting down while the terrain physics collider is in the middle of updating a heightfield, it was possible to reach the "wait.acquire()" without a running query, which meant that it would hang indefinitely. The fix is to verify that the terrain job was actually created successfully.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>