Fixed a race condition within AZ::IO::Streamer.
In rare cases it's possible for a read request to complete before the task thread issuing the request has had a change to update the processing states. In these cases FinalizeReads would not close out the read because it finds the request still waiting to be started. Eventually the task thread does set the processing status, but the scheduler thread won't be woken up. If this was the last read in a chain this can cause a deadlock waiting for the read to complete. This fix sets the state earlier but clears the OVERLAPPED structure so FinalizeReads doesn't complete requests that haven't been issued with the OS yet.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>