[Maestro] Fix crash at exit due to using an already destroyed allocator
This code uses global static instances, which have a nondeterministic
initialization order. Due to the fact that they are global, the AZ
allocators are not guaranteed to be initialized first. To deal with
this, most of this code already uses `std` containers over `AZStd` ones.
It stores a `CAnimPostFXNode::FxNodeDescriptionMap`, which is a
`std::map`, the values of which are `CFXNodeDescription`s. Those
descriptions were storing an `AZStd::string`. Putting an `AZStd::string`
in a global static variable is not guaranteed to work.

Signed-off-by: Chris Burel <burelc@amazon.com>