[Terrain] Macro materials can cover very large world areas (#9894)
Macro material information is stored in a grid across the terrain, with each tile in the grid supporting up to 4 macro materials and covering 64m x 64m. Previously the total number of indexable macro materials was limited to a uint16, which means it would break with worlds larger than 8km x 8km. Also each tile stored the full macro material for every entry at 32 bytes per entry / 128bytes per tile.

This change makes it so the total number of tiles is limited to uint32, which should be more than large enough for any amount of terrain with the current terrain system. This change also adds a level of indirection, so the macro materials themselves are stored in a separate buffer, and the tiles just reference that buffer with 16 bit indices. This means the 4 macro materials per tile only cost 2 bytes each / 8 bytes per tile.

______

* Macro materials now supported over a much larger world area. Added a level of indirection to avoid duplicating so much macro material data.

Signed-off-by: Ken Pruiksma <pruiksma@amazon.com>

* Minor bug fix

Signed-off-by: Ken Pruiksma <pruiksma@amazon.com>

* Updates from PR review suggestions

Signed-off-by: Ken Pruiksma <pruiksma@amazon.com>